% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoyeR-package.R
\docType{package}
\name{annoyeR-package}
\alias{annoyeR}
\alias{annoyeR-package}
\title{annoyeR: Approximate Nearest Neighbours (Oh Yeah) cpp11 wrapper}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nic Crane \email{thisisnic@gmail.com}

}
